fix(networkdata): stop race conditions causing exceptions when disconnecting aircraft
Previously, the task of deleting an aircraft would occur at roughly the same time as a task that
would be looking to assign them stands (for example) in response to the network data being updated.
Meaning that if the stand assignment was deleted, the stand assigner task could then come along and
try and re-add it, before which the aircraft itself could be deleted, causing a constraint
violation. This change moves the separate jobs required to disconnect an aircraft into a single job
that runs in a database transaction, running each of the previous jobs as subtasks. This stops the
race conditions and means that the deletion of other data such as stand assignments doesnt occur
until everything else goes with it. Events are also held onto until such a time that the commit
happens.