fix(regionalpressures): default negative regional pressures to zero
If the lowest QNH was 0, then the RPS would become -1 which would be a value violation as RPS are by
definition unsigned. This change will set the RPS to 0 if 0 is the lowest pressure (this shouldn't
be a thing, but apparently we get this from the metar feed now and again).

fix #712