srd: wrap srd imports in transactions
At the moment we truncate all SRD tables before we update. This is problematic in that if the update
process fails, the tables are wiped. This change fixes the issue by performing SRD updates in transactions.

We also make the SRD Routes primary key a bigint, so we don't have to reset the auto increment ids each time.